Bug fixes. Removed xor. New things.
First off, some bugs/issues have been fixed.
Necessary packages for multilib compiling
have been added to etc/install_deps.sh.

The PAM backdoor should be back to working
condition as of this commit. I'm not sure
about the accept() hook backdoor however.
Part of the problem was the read & write hooks.
I've excluded them from inc/includes.h, and I've
also disabled LOG_SSH and set it to be ignored
by default.
FILE_STEAL is also disabled for the time being.

Second, I've stripped the whole string xor'ing
system from the rootkit. I believe it was quite
inefficient and also unnecessary.

Third, I'm aware that the configuration stage
can be quite slow. So I've added support for
using pre-configured setups in the form of
gzip compressed tarballs. I've made a note
of this in the README, along with an example.

Other changes are just new functions and changes
to functions that were already in the rootkit.