Issues fixed & things tidied up
FILE_STEAL is operational again.
I just done a silly when it came to
iterating through the interesting_files
array. (namely, its size...)

I've added ssh.sh back to etc/ again.
It did used to live there, but I had
just kept it on my gist profile along
with a link to it within bdvl's README
so that people could download it if they
needed to, but I think practically it
makes more sense to just keep it here.

Other things include little changes to
the backdoor user's bashrc, such as a
colored prompt & only showing log msgs
when said log files exist.
I've also tidied up previous functions
within the rootkit, and removed const
definitions that weren't really necessary.

High up on my priorities atm is decreasing
the amount of time it can take to setup the
rootkit and its settings. I am currently
in the process of this.