Solved bedevil.sh taking ages to setup
Previously I had commented the fact that
I was unnecesarilly using a function within
bedevil.sh (settings.sh, really) once more
than really needed. I've fixed this.
Configuration/setup will now take a fraction
of the time that it previously took.

I've also done a little bit of work on the
way that the essential hooked function
arrays are written. See 'etc/arrays.sh' for
the changes and comments in relation to the
changes.

Considering the point above, the previous
'arrays' files are no longer valid. The only
file that gets read and written as C arrays
is libdl/hooks. This is due to the removal
of the whole string xoring system within bedevil.
Such a system for reading & writing arrays is
no longer necessary.

Lastly, I've changed some of the messages
that are shown during setup, and changed
how some of them are displayed.