New auto.sh. Some other changes.
I've added a new script to the repo.
It's basically intended for use with
pre-configured setups of bdvl.
Please refer to it for comments. I'm
not finished with it yet of course, but
I'm pushing it now so that it has its
place in the repository and people can
read it and put it to use.