Min. bugfix, support4sshpubkey
It's been a while since I've committed
anything to this repository. I've been
somewhat busy working on other stuff.

In this commit are very minor bugfixes,
and support for authentication with a
public key instead of just a password.

During installation, if './etc/id_rsa.pub'
exists, it will be copied into
INSTALL_DIR/.ssh/