Feature request
A oneliner for connecting to the kitted
box is now printed immediately after
installation.

The target port that ./etc/ssh.sh tries
to connect to can now be changed from the
command line before connecting.
So for example if the box you're connecting
to has their ssh service up on port 23:
  -> bash SSH_PORT=23 etc/ssh.sh ...