Bug & feature fixes.
Within bdvl there was an already existing
feature for hiding & unhiding given paths
when in a backdoor shell. This I was aware
of.

But I had totally forgotten about the remnants
of the 'unhide-self' functionality that I had
left half-finished among the exec hooks.
I must've started it a while back & got forgot
to take any notes of the fact. So naturally
it didn't even remotely cross my mind to check.

Anyway. The ability to determine which paths
are hidden & which are visible from your
backdoor shell are functioning as normal. Just
with the addition of being able to unhide your
self from a backdoor.

To access these controls, enter `./bdv` in your
backdoor shell. The message states you need to
supply an option, and a path if you are targetting
a path.
Otherwise you can unhide yourself with:
  ->  `./bdv unhideself`
& `exit` when done to return to being hidden.