Various fixes & changes
auto.sh now works much better than
what it did previously.

It has better error handling, in case
something should go wrong unexpectedly.

If it needs to create a file within a
directory that doesn't yet exist, auto.sh
will first create the directory before
trying to touch the file.

This was something that I overlooked as
being a problem, where the script will try
to create a file somewhere that doesn't
exist yet.

Additionally I've removed the accept hook
backdoor & removed references of it. It
was really just taking up space that wasn't
needed, especially considering that I hadn't
worked on it in about a year...

Last note in relation to changes is just
that some outputs have been changed or added
to better suit what is going on.