Additions & removals
The BD_README that would normally
appear on first login has been removed
from the repo & the kit. There was just
no need for it. If it contained useful
information, sure, there might've been
a good purpose for it, but imo it said
nothing that the repo's README doesn't
already say, and having that show up on
every new installation is just an annoyance.

The big ascii that would normally show
upon backdoor login is now also no longer.
Instead of this, a random quote from
~/.rolf is shown. The parts of the rootkit
responsible for making sure this file is
copied have been updated accordingly.

The 'patch_sshdconfig' function is now
present in auto.sh but is by default not
run. The line that would run it is commented
out. So you can either uncomment the line
or move it to wherever else you think is
more suitable.

The main function that determines whether
or not a specified port is that of a hidden
port now only uses 13 bytes for the buffer
per line of hide_ports, compared to the
previous 128. That was just an unnecessary
size for such small input. There is absolutely
no need for any size bigger than 13 bytes.