Dynamic magic GID & rootkit sshdpatch
There are a considerable amount of changes
in this commit.

The most notable changes being:
  a. the ability to change the rootkit's
     magic GID from within a backdoor shell.
  b. the rootkit will now (if instructed to do so)
     make sure the sshd_config file is up to par
     in order to always allow for backdoor access.
     as opposed to the setup scripts doing this once
     and that being it...
I have updated the README to make note of these
changes. The source code for both additions has
been commented plenty.

Due to the addition of point b I have removed
the functions & references of them in the setup
scripts that would've usually done this.

The addition of point a means that all functions
which need to know the magic GID now need to read
it from the file beforehand.

If READ_GID_FROM_FILE is either disabled, or its
responsible functions fail to read the GID, the
original (MAGIC_GID) value is returned instead.
Perhaps this isn't the optimal choice. It is for now.

The backdoor bashrc will force hide all files
within the installation directory upon logging
in. This just triple enforces that all files
are hidden & STAY hidden.

I removed etc/names.txt as nothing actually uses or
needs it anymore.
dialog.sh had some value changes to fit more
settings on the screen at one time.

Other changes are just fixes really.