Automatic GID changing & other stuff
GID:
Most notable change in this commit is the
addition of the rootkit's ability to refresh
its magic GID based on a interval.
Due to this I've also created a new directory
solely for functionality relating to the rootkit's
GID. Comments in abundance.

I made small changes to the way manual changing
of the GID works from within the backdoor shell
as well.

OTHER:
Additional preprocessor checks just to make sure
settings that are enabled are sane.

Fixed nonsensical logic. To.... well make more
sense.

Random paths that the rootkit will use are a bit
more up to par in terms of seeming like standard
paths. i.e.: the rootkit will not need to create
             directories that do not yet exist.
             it will just set stuff up in the
             root of the directories instead of
             trying to go a directory deeper.
             this just seems safer & less weird..