Automatic maintaining of external hidden paths
Upon the addition of READ_GID_FROM_PATH, I had
commented the fact that I did not really have
a sane way of keeping track of files created
outside of the rootkit's installation directory
in a backdoor shell/by a general rootkit process.

I have created a system for keeping track of
files created by the rootkit's magic user.
If it's sane or not that's yet to be seen.
The README has been updated briefly to explain
a little how this functionality works.

Basically, files you create in backdoor shell
are kept track of in a file in install dir.
Paths in that file that have been unhidden
post-creation will be ignored when changing
magic GID.