Improvements. General updates & fixes.
bdvl has now the functionality in it
where it will make sure that the
backdoor .bashrc & .profile stay up
to par so that the backdoor works as
intended. i.e.: if one of them is
                accidentally removed
                they will both be
                rewritten.

in the past, when making paths for
stuff, links would be made just once
before preloading the rootkit.
now the rootkit will handle doing
this linking of paths into the
installation directory.
this is done with `./bdv makelinks`.

the HIDE_MY_ASS functionality is
a tad smarter. in terms of knowing
fully what it does not need to
keep track of, as that stuff is
technically always being kept track
of by the rootkit.

in general I've just tried to make
everything that bit more efficient
& smarter.