General upgrade from previous bdvl...
A lot has been added & changed
in this compared to the contents
of the previous.

Right off the bat, there is a new
method for sshd patching, dubbed
SOFT_PATCH_SSHD_CONFIG appropriately.
It applies the exact same logic as
the original method but instead
only applies to sshd's read of the
file. No direct file writing
necessary for this method.

HIDE_MY_ASS is MUCH smarter than
it previously was. Hidden directories
are kept on top of now.
Is also accompanied by UNINSTALL_MY_ASS
which, when doing `./bdv uninstall`
will recursively remove all of the
hidden directories kept track of by
HIDE_MY_ASS.

The actual installation process
is totally different now. auto.sh
does very little in comparison now.
The rootkit solely is responsible for
all installation & patching of the
dynamic linker. Since the rootkit is
now responsible for patching the
dynamic linker, when doing './bdv uninstall'
it will now revert the dynamic linker
preload path back to the original.
(/etc/ld.so.preload)
All auto.sh now needs to know is
BDVLSO..

The file stealing utility is now
much better in general.
Files are now stolen in chunks, and
the settings relating to this
can be found in setup.py.
Since files are now stolen in chunks,
I raised the default limit for
files to steal to 20mb.
I also brought back the symlink-fallback
option to bdvl. If a file exceeds the
MAX_FILE_SIZE value, or there is an
error copying the contents of the file,
a link for it will be made instead..

There are a list of ports in setup.py
that bdvl will always hide. (they can't
be unhidden) PAM_PORT is right now the
only one.

There is a lot of other stuff but off
of the top of my head I don't exactly
remember everything...