New backdoor methods. Various improvements...
There is now an accept hook backdoor
back in bdvl. This was in an older
version of bdvl but I had since
removed it. So now it is back.

The new backdoor addition is the
ICMP backdoor. The rootkit spawns
a hidden process for this backdoor.

For more on how these work go see
inc/backdoor/*...
There are plenty of comments in
the README & in setup.py for these
backdoors.

A handful of various minor issues
have been resolved.

The dynamic linker patching
part of the rootkit will now
automatically search for valid
ld.so paths on the system, instead
of having to manually feed them
to setup.py beforehand.

There is now an exec logger in
the kit. Everything that users
execute after installation is
logged. The kit already will
steal shell history files, but
even if they do not have a
history file, stuff they do
will be logged...

There are some other minor
changes that I don't particularly
remember but everything noted
above is all of the important
stuff...

Oh, very last thing I suppose...
HIDE_MY_ASS didn't work
properly due to a change I had
made & forgot to fix up in this
repository. So now it's fixed.