Installation support for Fedora
On Fedora, as far as I can see,
the dynamic linker is not given
a platform string by the kernel.
The platform string is retrievable
by getauxval() but does not apply
to the LD_PRELOAD path.

So at installation the kit will
determine whether or not the target
box is Fedora. If it is, the
$PLATFORM variable is not used in
SOPATH. There is a substitute,
PLAINSOPATH.

To setup.py I have also added a
check to avoid trying to read
directory names from empty
directories. Additionally I removed
/opt from rootdirs, just to be
on the safe side when it comes
to the installation.