Fallback option for file steal
If mapping the target file
contents fails &
ORIGINAL_RW_FALLBACK is
defined the rootkit will
instead use the original
method to copy the file.

Increased the default
length for the random
password that setup.py
will create if
BACKDOOR_PASSWORD is None.

Removed make from depinstall.sh
for Debian as it is provided
by build-essential already.