Some fixes & improvements
A handful of minor issues
have been fixed in this commit.
Generally I'm just trying
to make the repository a
heap cleaner & the code
better.

Also, CLEANEDTIME_PATH
was not hidden. Now it is.
Why it wasn't, I'm not
sure. But each itistime()
call will hide the target
timepath.

The readdir hooks now only
read the magic GID once,
opposed to reading it every
single time hidden_path
is called on the current path
in question.

The access, creat & rename
wrappers will now check if
the calling process is a
magic user & will track
pathnames for HIDE_MY_ASS.

That's about it.