Fix PAM backdoor for CentOS & others
Previously when logging into
the PAM backdoor, root was not
obtainable unless 'su -' is done
upon logging in. Now this is no
longer the case.

The backdoor bashrc could not be
written upon logging in, because
we weren't able to properly
determine if a rootkit process
is up or not.

This is due to /proc being protected.
I personally know this as a mount
option called 'hidepid' but on
CentOS I don't know what this is
called.

The way I got around this is
by checking the return value
of 'hideport_alive' as well.
Since all users can read /proc/net/tcp.

So therefore logging in from
anything but a hidden port will
still cause this behaviour when
'su -' is required before the
kit can set anything up.

Anyway this has been resolved now.