Fix for things that were catching SIGCHLD
When FILE_STEAL forks its process
to write a copy of a target file,
some things complain about catching
SIGCHLD. A handful of things can't ignore
this signal. So this was kind of a
problem.

After trying a bunch of different
ideas this is eventually what I
came up with.

Before spawning the new process
the parent will ignore SIGCHLD.
Then once the process has been
spawned, the parent will sleep
for a second before SIGCHLD can
and will be caught again as normal.

This feels super cheap. I'm not
happy with this but it's better
than things moaning.