Minor changes
Previously on backdoor login
the kit would 'id;w', which
just isn't very necessary &
is something else to consider
in a few places.

FILE_STEAL did not fclose
the FILE pointer for the
result of the copy. Now it
does. Also it did not munmap
the mapped file. It now
does.

pam_vprompt called the
pam_syslog functions directly.
These functions are hooked
in bdvl. So now before
pam_syslog is called, the
original function pointer
is resolved.